// src/views/Lists/Articules/interfaces.ts
export interface ModalTypes {
    KALITEID?: string;
    KALITEADI?: string;
    HAMID?: string;
    HAMADI?: string;
    SIPARISPROSESI?: number;
    ENISTENEN?: string;
    GRAMAJISTENEN?: string;
    PUS?: string;
    FEINE?: string;
    IPLIKBOYU1?: string;
    IPLIKBOYU2?: string;
    DOVIZID?: string;
    COMPOSITION?: string;
    ORMETOPKILOSU?: string;
    KAPALI?: string;
    ARTICULTIPIID?: string;
    ARTICULTIPIADI?: string;
    KALINLIKID?: string;
    KALINLIKADI?: string;
    IPLIKCINSIID?: string;
    IPLIKCINSIADI?: string;
    VARSAYILANBIRIMADI?: string;
}


export interface IModalForm {
  KALITEID?: string;
  KALITEADI?: string;
  HAMID?: string;
  HAMADI?: string;
  SIPARISPROSESI?: number;
  ENISTENEN?: string;
  GRAMAJISTENEN?: string;
  PUS?: string;
  DOVIZID?: string;
  COMPOSITION?: string;
  ORMETOPKILOSU?: string;
  KAPALI?: string;
  MAYLI?: string;
  IPLIKBOYU1?: string;
  IPLIKBOYU2?: string;
  ARTICULTIPIID?: string;
  ARTICULTIPIADI?: string;
  KALINLIKID?: string;
  IPLIKCINSIID?: string;
  MAXFIREORANI?: string;
  MINFIREORANI?: string;
  BIRIMID?: string;
  NOTU?: string;
  MAXIMUMFIREORANI?: string;
  MINIMUMFIREORANI?: string;
  FEINE: number;
  COMPOSITION_RU?: string;
  COMPOSITION_ENG?: string;
} 